# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "ETH Fee Calculator",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/transactions/cost-per-hour": {
            "get": {
                "description": "Get transaction costs per hour",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction costs per hour",
                "operationId": "get-transaction-per-hour",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Initial period to fetch the data",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Final period to fetch the data",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.TransactionCostPerHourResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "presenter.TransactionCostPerHourResponse": {
            "type": "object",
            "properties": {
                "t": {
                    "type": "integer"
                },
                "v": {
                    "type": "number"
                }
            }
        }
    }
}